getBBox() on a SVGPathElement with curves incorrectly includes control points https://bugs.webkit.org/show_bug.cgi?id=53512 <rdar://problem/9861154>
Reviewed by Dirk Schulze.

The CoreGraphics implementation of Path::boundingRect() called
CGPathGetBoundingBox, which includes the path's control points in its
calculations. Snow Leopard added CGPathGetPathBoundingBox, which
finds the bounding box of only points within the path, and does not
include control points. On Snow Leopard and above, we now use the latter.


Identifier: 83044@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@94116 268f45cc-cd09-0410-ab3c-d52691b4dbfc