Group failing tests in garden-o-matic if there are more than three in a folder https://bugs.webkit.org/show_bug.cgi?id=67373
Reviewed by Dimitri Glazkov.

Sometimes a failure will cause many tests in a given directory to fail.
When that happens, it's better to list the directory as a single item
(with a count) rather than spamming all the test names.

(This patch also updates a bunch of unit test expected results after my last commit.)

* BuildSlaveSupport/build.webkit.org-config/public_html/TestFailures/scripts/base.js:
* BuildSlaveSupport/build.webkit.org-config/public_html/TestFailures/scripts/base_unittests.js:
* BuildSlaveSupport/build.webkit.org-config/public_html/TestFailures/scripts/ui/failures_unittests.js:
* BuildSlaveSupport/build.webkit.org-config/public_html/TestFailures/scripts/ui/notifications.js:
* BuildSlaveSupport/build.webkit.org-config/public_html/TestFailures/scripts/ui/notifications_unittests.js:


Identifier: 83218@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@94309 268f45cc-cd09-0410-ab3c-d52691b4dbfc