Changes to aria-hidden don't change VO navigation https://bugs.webkit.org/show_bug.cgi?id=67722
Reviewed by Darin Adler.

Source/WebCore:

When altering aria-hidden, WebCore needs to update the children caches of
affected elements. However, for elements that were children, but ignored, their
caches did not get updated, and stale information would be propagated.

The fix is to always clearChildren() when a parent is asking for children. This
ensures information is always up to date when the parent itself is asking for new data.

Test: accessibility/aria-hidden-updates-alldescendants.html

* accessibility/AccessibilityObject.h:
* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::ariaIsHidden):
(WebCore::AccessibilityRenderObject::childrenChanged):
(WebCore::AccessibilityRenderObject::addChildren):

LayoutTests:

* accessibility/aria-hidden-updates-alldescendants-expected.txt: Added.
* accessibility/aria-hidden-updates-alldescendants.html: Added.



Identifier: 83611@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@94716 268f45cc-cd09-0410-ab3c-d52691b4dbfc