offsetFromRoot optimization is disabled after r93837 https://bugs.webkit.org/show_bug.cgi?id=67677
Reviewed by Simon Fraser.

Unfortunately not covered by any test as it does not change updateLayerPosition complexity (it is still O(n^2)
because of clippedOverflowRectForRepaint). However it is a noticeable slowdown on some use cases but we don't
have a good test harness to cover such slowdown.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateLayerPositions): Pass offsetFromRoot as not doing so means that we miss
the optimization altogether.

(WebCore::RenderLayer::computeRepaintRects):
(WebCore::RenderLayer::updateLayerPositionsAfterScroll):
* rendering/RenderLayer.h:
Renamed |cachedOffset| to |offsetFromRoot| to be consistent with our latest renaming.

Identifier: 83636@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@94741 268f45cc-cd09-0410-ab3c-d52691b4dbfc