https://bugs.webkit.org/show_bug.cgi?id=67898 REGRESSION(r94900): fast/images/support-broken-image-delegate.html fails on Mac
Reviewed by Simon Fraser.

Source/WebCore:

New function willPaintBrokenImage() returns true when there has been an error
loading the image and the broken image icon will be used in its place. This is
necessary since it is possible to have an error loading an image and to NOT use
the broken image icon.
* loader/cache/CachedImage.cpp:
(WebCore::CachedImage::willPaintBrokenImage):
* loader/cache/CachedImage.h:
* rendering/RenderImage.cpp:
(WebCore::RenderImage::imageSizeForError):

LayoutTests:

* platform/mac/Skipped:



Identifier: 83865@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@94980 268f45cc-cd09-0410-ab3c-d52691b4dbfc