REGRESSION (64275): Shape pattern-image fill turns black https://bugs.webkit.org/show_bug.cgi?id=51061 <rdar://problem/8504705>
Reviewed by Simon Fraser.

Source/WebCore:

When destroying a resource, register clients who are losing their
resource as having pending resources, so they can be resolved in the case a
resource with that id is re-registered.

Test: svg/custom/pending-resource-after-removal.xhtml

* rendering/svg/SVGResourcesCache.cpp:
(WebCore::SVGResourcesCache::resourceDestroyed):
* svg/SVGStyledElement.h:
(WebCore::toSVGStyledElement): Added.

LayoutTests:

Add a test ensuring that replacing a SVG pattern without changing
the id causes the clients to be updated properly.

* svg/custom/pending-resource-after-removal-expected.png: Added.
* svg/custom/pending-resource-after-removal-expected.txt: Added.
* svg/custom/pending-resource-after-removal.xhtml: Added.


Identifier: 83930@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95047 268f45cc-cd09-0410-ab3c-d52691b4dbfc