WebCore::Cursor::ensurePlatformCursor() should always set a valid platform cursor on Windows https://bugs.webkit.org/show_bug.cgi?id=68043
Make sure we set a valid platform cursor in the Cursor::NoDrop case, and add a default
clause that uses the arrow cursor in case another cursor type is added in the future
and we forget to update ensurePlatformCursor().

Reviewed by Anders Carlsson.

* platform/win/CursorWin.cpp:
(WebCore::Cursor::ensurePlatformCursor): Always set a valid platform cursor.


Identifier: 83940@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95057 268f45cc-cd09-0410-ab3c-d52691b4dbfc