Port MachineStackMarker to Windows ARM and MIPS https://bugs.webkit.org/show_bug.cgi?id=68068
Reviewed by Geoffrey Garen.

Use the correct memeber of the CONTEXT struct for the stackpointer for CPU(ARM) and CPU(MIPS).
Only query CONTEXT_INTEGER and CONTEXT_CONTROL, since CONTEXT_SEGMENTS isn't defined for
CPU(ARM) and CPU(MIPS) and the stackpointer is defined in the CONTEXT_CONTROL section for
CPU(ARM), CPU(X86) and CPU(X86_64) and in the CONTEXT_INTEGER section for CPU(MIPS).

* heap/MachineStackMarker.cpp:
(JSC::getPlatformThreadRegisters):
(JSC::otherThreadStackPointer):


Identifier: 84035@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95153 268f45cc-cd09-0410-ab3c-d52691b4dbfc