[n]stricteq code is bogus in JSValue32_64 JIT https://bugs.webkit.org/show_bug.cgi?id=68141
Reviewed by Sam Weinig.

The code tries to check for both ints or cells, but this check also
catches cases where values that are undefined, null, etc (probably
was incorrectly assuming cell was the 2nd highest tag?).

Also, there is no need not to handle int on the fast path.
stricteq is just a case of comparing the payloads, if we:
    * handle cases of differing tags on a slow path
    * handle doubles a slow path
    * handle both-are-string on a slow path

* jit/JITOpcodes32_64.cpp:
(JSC::JIT::compileOpStrictEq):
(JSC::JIT::emitSlow_op_stricteq):
(JSC::JIT::emitSlow_op_nstricteq):



Identifier: 84050@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95168 268f45cc-cd09-0410-ab3c-d52691b4dbfc