Experiment with removing ability to call a collection (except document.all) https://bugs.webkit.org/show_bug.cgi?id=67579
Reviewed by Anders Carlsson.

Source/WebCore:

At the request of the public-script-coord mailing list (specifically Brendan Eich, see
http://lists.w3.org/Archives/Public/public-script-coord/2011JulSep/0360.html), this
patch removes the ability to call a collection (either a NodeList or HTMLCollection,
but not an HTMLAllCollection) as function, a syntax that we adopted to emulate IE.
It is being landed to find out if there are any sites relying on this behavior of WebKit,
or, if it is only used in IE only paths.  If we find sites are breaking, it should be rolled
out and we should inform the public-script-coord mailing list.

* bindings/js/JSHTMLAllCollectionCustom.cpp:
Update comment.

* bindings/js/JSHTMLCollectionCustom.cpp:
* bindings/js/JSNodeListCustom.cpp:
Remove custom call code.

* bindings/scripts/CodeGeneratorV8.pm:
Add support for V8CustomCall.

* dom/NodeList.idl:
* html/HTMLCollection.idl:
Remove CustomCall.

LayoutTests:

* fast/dom/Element/id-in-formcollection-expected.txt:
* fast/dom/Element/id-in-formcollection.html:
* fast/dom/HTMLOptionElement/collection-setter-getter-expected.txt:
* fast/dom/HTMLOptionElement/collection-setter-getter.html:
Don't use call syntax for tests that aren't explicitly testing it.

* fast/dom/NodeList/nodelist-item-call-as-function-expected.txt:
* fast/dom/NodeList/script-tests/nodelist-item-call-as-function.js:
Update test to show that we throw on call.

* fast/profiler/call-nodelist-as-function-expected.txt: Removed.
* fast/profiler/call-nodelist-as-function.html: Removed.
Remove test of removed feature.


Identifier: 84084@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95203 268f45cc-cd09-0410-ab3c-d52691b4dbfc